#302 refactored web interface architecture
