+++ interfaces/web/static/js/common/bot_connection.js
@@ -1,0 +1,84 @@
+// utility variables
+var config_key_attr = "config-key";
+var config_value_attr = "config-value";
+var update_url_attr = "update-url";
+var config_element_class = "config-element";
+
+// utility functions
+function log(text){
+    window.console&&console.log(text);
+}
+
+function get_update(){
+    $.ajax({
+      url: "/update",
+    }).done(function(data) {
+        unlock_ui();
+
+    }).fail(function(data) {
+        lock_ui();
+
+    }).always(function(data) {
+        manage_alert(data);
+  });
+}
+
+function manage_alert(raw_data){
+    data = JSON.parse(raw_data)
+    $.each(data, function(i, item) {
+        create_alert(data[i].Level, data[i].Title, data[i].Message);
+    })
+}
+
+function handle_route_button(){
+    $(".btn").click(function(){
+        button = $(this)
+        if (button[0].hasAttribute('route')){
+            command = button.attr('route');
+            origin_val = button.text();
+            $.ajax({
+                url: command,
+                beforeSend: function() {
+                    button.html("<i class='fas fa-circle-notch fa-spin'></i>");
+                },
+                complete: function() {
+                   button.html(origin_val);
+                }
+            });
+         }
+    });
+}
+
+function send_and_interpret_bot_update(updated_data, update_url, dom_root_element, success_callback, error_callback){
+    $.ajax({
+        url: update_url,
+        type: "POST",
+        dataType: "json",
+        contentType: 'application/json',
+        data: JSON.stringify(updated_data),
+        success: function(msg, status){
+            if(typeof success_callback === "undefined") {
+                update_dom(dom_root_element, msg);
+            }
+            else{
+                success_callback(updated_data, update_url, dom_root_element, msg, status)
+            }
+        },
+        error: function(result, status, error){
+            window.console&&console.error(result);
+            window.console&&console.error(status);
+            window.console&&console.error(error);
+            if(typeof error_callback === "undefined") {
+                create_alert("danger", "Error when handling action: "+result.responseText+".", "");
+            }
+            else{
+                error_callback(updated_data, update_url, dom_root_element, result, status, error)
+            }
+        },
+    })
+}
+
+$(document).ready(function () {
+    handle_route_button();
+    setInterval(function(){ get_update(); }, 500);
+});
