+++ interfaces/web/templates/advanced_index.html
@@ -1,0 +1,8 @@
+{% extends "layout.html" %}
+{% block body %}
+<br>
+<div class="alert alert-danger" role="alert">
+    <h1>This interface is currently in development</h1>
+</div>
+
+{% endblock %}
