+++ interfaces/web/templates/config.html
@@ -78,4 +78,5 @@
     <br>
     <button type="button" class="btn btn-primary">Save config</button>
 </div>
+<script src="{{ url_for('static', filename='js/components/configuration.js') }}"></script>
 {% endblock %}
